Revert "Merge branch 'master' into fix-synchronize-libcontroller"
This reverts commit 09940ed73003dad14987c3849415d1afc05f856e, reversing
changes made to f731f86ba737fe0a46725f6b2cd31a1feb832eff.

+++ webots_ros2_control/include/webots_ros2_control/Ros2Control.hpp
@@ -47,8 +47,8 @@ namespace webots_ros2_control
       webots_ros2_driver::WebotsNode *mNode;
       std::shared_ptr<pluginlib::ClassLoader<Ros2ControlSystemInterface>> mHardwareLoader;
       std::shared_ptr<controller_manager::ControllerManager> mControllerManager;
-      int mControlPeriodMs;
-      int mLastControlUpdateMs;
+      double mControlPeriodMs;
+      double mLastControlUpdateMs;
 
       std::thread mThreadExecutor;
       rclcpp::executors::MultiThreadedExecutor::SharedPtr mExecutor;
